Deduplicate submit button logic using custom hook
