+++ frontend/packages/@depmap/dataset-manager/utils/disableSubmitButton.ts
@@ -1,0 +1,14 @@
+export function useSubmitButtonIsDisabled(
+  requiredProperties: string[] | undefined,
+  formData: any
+) {
+  if (requiredProperties !== undefined) {
+    const requiredFormValues = requiredProperties.map((prop) => {
+      return formData[prop];
+    });
+    return !requiredFormValues.every((val) => {
+      return val !== undefined && val !== null;
+    });
+  }
+  return false;
+}
